Add level flight AoA marker for dry craft
Currently, the only level flight AoA marker present isn't very useful in most cases because by the time you get to a point in an ascent profile, the craft's mass is most likely different from the current mass, requiring constant fuel tank tweaking to identify behavior at specific points. Providing an additional marker for empty craft highlights to the user the region a craft will most likely be used in allowing faster behavior assessment, as well as avoiding having to empty all tanks in order to check how a craft behaves on landings. 

In order to distinguish between both markers, they've been colored with their respective torque chart colors. This could potentially be seen as a bug to some users thinking the vertical lines are a singularity on the torque charts, but i don't quite have an idea of how to best avoid this issue. This pull request also changes mouseover label on the vertical line tooltip to provide additional info but that may not be enough.